♻️ refactor(alarm): 알람 컴포넌트 구조 재설계 및 타입 안정성 강화
- 알람 관련 UI 컴포넌트(`AlarmItem`, `AlarmTopBar` 등)를 `component` 패키지로 이동 및 통합
- 홈 상단 바 관련 컴포넌트 위치를 `ui.home.bar` 패키지로 변경하여 구조 정리
- `AlarmRepository`, `ViewModel`, `PagingSource`에서 알람 타입을 `String` 대신 `AlarmType` Enum을 직접 사용하도록 리팩터링
- `AlarmType` 아이콘 리소스 이름 오타 수정 (`ic_quration` -> `ic_curation_alarm`) 및 확장 프로퍼티 업데이트
- `AlarmFilterTabs`에서 하드코딩된 리스트 대신 `AlarmType.entries`를 사용하도록 개선
- `AlarmScreen`의 선택된 탭 상태 유지에 `rememberSaveable` 적용
- `AIArticleModal` 컴포넌트의 프리뷰 코드 추가
- `data` 모듈에서 불필요한 `paging-compose` 의존성 제거 및 `AlarmPagingSource` 로직 단순화
