+++ core/src/main/java/com/linku/core/repository/AlarmRepository.kt
@@ -1,11 +1,12 @@
 package com.linku.core.repository
 
 import com.linku.core.model.alarm.AlarmList
+import com.linku.core.model.alarm.AlarmType
 
 interface AlarmRepository {
 
     suspend fun fetchAlarms(
-        type: String,
+        type: AlarmType,
         cursor: Long?,
         size: Int
     ): Result<AlarmList>
