+++ data/build.gradle.kts
@@ -104,6 +104,5 @@ dependencies {
     implementation(libs.kotlinx.serialization.json)
 
     // paging3
-    implementation(libs.paging.compose)
     implementation(libs.paging.runtime)
 }
