+++ feature/home/src/main/java/com/linku/home/ui/alarm/component/AlarmNothingTab.kt
@@ -1,4 +1,4 @@
-package com.linku.home.ui.alarm.top.bar.component
+package com.linku.home.ui.alarm.component
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
