+++ feature/home/src/main/java/com/linku/home/ui/home/bar/component/SelectedSummaryRow.kt
@@ -1,4 +1,4 @@
-package com.linku.home.ui.top.bar.component
+package com.linku.home.ui.home.bar.component
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
